+++ extensions/textline/actions/setContactConsent/config/index.ts
@@ -1,0 +1,1 @@
+export { FieldsValidationSchema, fields } from './fields'