+++ extensions/textline/actions/setContactConsent/setContactConsent.ts
@@ -1,0 +1,69 @@
+import { z, ZodError } from 'zod'
+import { fromZodError } from 'zod-validation-error'
+import { type Action } from '@awell-health/extensions-core'
+import { type settings, SettingsValidationSchema } from '../../settings'
+import { Category, validate } from '@awell-health/extensions-core'
+import { FieldsValidationSchema, fields } from './config'
+import TextLineApi from '../../client/textLineApi'
+
+export const setContactConsent: Action<typeof fields, typeof settings> = {
+  key: 'setContactConsent',
+  title: 'Set Contact Consent',
+  description:
+    'Set contact consent status. You can set consent as true or false. ',
+  category: Category.COMMUNICATION,
+  fields,
+  previewable: true,
+  onActivityCreated: async (payload, onComplete, onError) => {
+    try {
+      const {
+        settings: { accessToken },
+        fields: { recipient, consentStatus },
+      } = validate({
+        schema: z.object({
+          settings: SettingsValidationSchema,
+          fields: FieldsValidationSchema,
+        }),
+        payload,
+      })
+
+      const textLineApi = new TextLineApi(accessToken)
+      await textLineApi.setContactConsent(
+        consentStatus,
+        recipient,
+      )
+
+      await onComplete()
+    } catch (err) {
+      if (err instanceof ZodError) {
+        const error = fromZodError(err)
+        await onError({
+          events: [
+            {
+              date: new Date().toISOString(),
+              text: { en: error.message },
+              error: {
+                category: 'BAD_REQUEST',
+                message: error.message,
+              },
+            },
+          ],
+        })
+      } else {
+        const message = (err as Error).message
+        await onError({
+          events: [
+            {
+              date: new Date().toISOString(),
+              text: { en: message },
+              error: {
+                category: 'SERVER_ERROR',
+                message,
+              },
+            },
+          ],
+        })
+      }
+    }
+  },
+}