+++ extensions/textline/CHANGELOG.md
@@ -1,5 +1,10 @@
 # TextLine changelog
 
+## 2024-03-5
+
+1. Add capability to set contact consent status (https://textline.docs.apiary.io/#reference/customers/set-contact-consent/set-contact-consent)
+2. Add several fixes and tweaks to improve current actions
+
 ## 2024-02-26
 
 1. Add capability to search for messages using all the allowed filters provided by TextLine API.